Add AuthenticationResponse and PasswordChallengeResponse to UserManagement resources
